Fix FolderMetadata
`decodeListResponse` decodes `FolderMetadata` records as
`DeletedMetadata`. This causes `ListFolderResponse.entries` to contain
only `FileMetadata` and `DeletedMetadata`; entries that should be
`FolderMetadata` are instead decoded as `DeletedMetadata`.

The cause was that `FolderMetadata.sharingInfo` had type
`FileSharingInfo`. This type has different fields than the Dropbox API's
FolderMetadata's `folder_sharing_info`.

This commit introduces `FolderSharingInfo`, and changes the type of
`FolderMetadata.sharingInfo` to this instead.

The error was hidden by the fact that `decodeListResponse` ignores the
JSON `.tag` property, and instead uses the first decoder that succeeds.
Since `decodeFolderMetadata` always failed, folder metadata always fell
through to `decodeDeletedMetadata`, which ignored the `.tag` and the
extra fields and succeeded.

This commit modifies `decodeListResponse` to use `Json.Decode.union`
instead of `Json.Decode.oneOf`. This insures that decoding errors in the
file and folder tag case are reported as such, instead of decoded as
`DeletedMeta`.

I verified the behavior in the context of an application that uses the
library, and with live data: by modifying `decodeListResponse` and
verifying that the application reported a decoding error; then by
adding `FolderSharingInfo`, and verifying that the application's
display of Dropbox data matched the file hierarchy presented by the
Dropbox web interface. (Previous to these changes it did not.)